+++ gui/ta-gui/src/app/aluno.service.ts
@@ -1,0 +1,8 @@
+import { Aluno } from './aluno';
+
+export class AlunoService {
+  alunos: Aluno[] = [];
+  gravar(aluno: Aluno): Aluno {
+     this.alunos.push(aluno);
+  }
+}
