+++ gui/ta-gui/src/app/aluno.ts
@@ -1,0 +1,6 @@
+export class Aluno {
+  nome: string;
+  cpf: string;
+  email: string;
+  github: string;
+}
