+++ gui/ta-gui/src/app/app.component.html
@@ -10,6 +10,7 @@
 <tr>
    <td><label>Nome </label></td>
    <td><input [(ngModel)]="aluno.nome" ></td>
+   <td><button (click)="gravar(aluno)">Adicionar</button></td>
 </tr>
 <tr>
    <td><label>CPF </label></td>
