+++ gui/ta-gui/src/app/app.component.ts
@@ -1,17 +1,21 @@
 import { Component } from '@angular/core';
 import { NgModule } from '@angular/core';
 
+import { Aluno } from './aluno';
+import { AlunoService } from './aluno.service';
+
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
+
 export class AppComponent {
    aluno: Aluno = {nome: "", cpf: "", email: "", github: ""};
-}
-export class Aluno {
-  nome: string;
-  cpf: string;
-  email: string;
-  github: string;
+   alunoService = new AlunoService();
+
+   gravar(a: Aluno): void {
+     this.alunoService.gravar(a);
+     this.aluno = {nome: "", cpf: "", email: "", github: ""};
+  }
 }
